# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
        "url": "/"
    }],
...
    "paths": {
...
        "/admin/users/all": {
            "get": {
                "parameters": [{
                        "name": "locataire",
                        "in": "query",
                        "required": false,
                        "description": "true ou false pour l'activation du filtre",
                        "type": "bool"
                    },
                    {
                        "name": "proprietaire",
                        "in": "query",
                        "required": false,
                        "description": "true ou false pour l'activation du filtre",
                        "type": "bool"
                    }, {
                        "name": "demarcheur",
                        "in": "query",
                        "required": false,
                        "description": "true ou false pour l'activation du filtre",
                        "type": "bool"
                    }
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Admin request"
                ],
                "summary": "Permet de voir la liste des utilisateurs",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/components/schemas/users"
                        }
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
        "/admin/apparts/all": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Admin request"
                ],
                "summary": "Permet de voir la liste des appartements",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
        "/admin/appart/certificate": {
            "post": {
                "tags": [
                    "Admin request"
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "summary": "Certified  an appart",
                "requestBody": {
                    "description": "certified an appart",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/certificate_appart"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/certificate_appart"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "certificate_appart": {
                "type": "object",
                "properties": {
                    "appart_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "certified": {
                        "type": "bool",
                        "format": "bool",
                        "example": true
                    }
                }
            },
...
            "users": {
                "type": "object",
                "properties": {
                    "locataire": {
                        "type": "bool",
                        "format": "bool",
                        "example": true
                    },
                    "proprietaire": {
                        "type": "bool",
                        "format": "bool",
                        "example": true
                    },
                    "demarcheur": {
                        "type": "bool",
                        "format": "bool",
                        "example": true
                    }
                }
            }
        }
    }
}